update NotificationSelector so it adds user and restaurant name in method

+++ P2/restaurants/views/FollowView.py
@@ -43,12 +43,13 @@ class FollowUnfollowView(APIView):
         if request.user.is_authenticated:
             if not Restaurant.objects.filter(id=self.kwargs['rest_id']).exists():
                 return Response({'Error': "Restaurant not Found"}, status=status.HTTP_404_NOT_FOUND)
+            if not Following.objects.filter(user=self.request.user,
+                                            restaurant=Restaurant.objects.get(id=self.kwargs['rest_id'])).exists():
+                return Response({'Success': "User never followed restaurant."}, status=status.HTTP_200_OK)
 
-            try:
-                Following.objects.filter(user=self.request.user,
-                                         restaurant=Restaurant.objects.get(id=self.kwargs['rest_id'])).delete()
-                return Response({'Success': "User does not follow restaurant anymore."}, status=status.HTTP_200_OK)
-            except:
-                return Response({'Error': "User Cannot UnFollow"}, status=status.HTTP_400_BAD_REQUEST)
+            # Delete Following of Restaurant
+            Following.objects.filter(user=self.request.user,
+                                     restaurant=Restaurant.objects.get(id=self.kwargs['rest_id'])).delete()
+            return Response({'Success': "User does not follow restaurant anymore."}, status=status.HTTP_200_OK)
 
         return Response({'Error': "User not Authenticated"}, status=status.HTTP_401_UNAUTHORIZED)
